wrap game page in server component for initial data fetch
