+++ client/src/app/game/[code]/not-found.tsx
@@ -1,0 +1,8 @@
+export default function NotFound() {
+  return (
+    <div className="flex flex-col items-center justify-center gap-4">
+      <div className="text-2xl font-bold">Error</div>
+      <div className="text-xl">Game not found</div>
+    </div>
+  );
+}
